notebooks/GenAI/requirements.txt
python-dotenv
openai 
pandas 
numpy 
streamlit
azure-search-documents
azure-identity
